added more apis

+++ routes/api.php
@@ -18,10 +18,15 @@ use Illuminate\Http\Request;
 // });
 
 //api routes
-Route::get('/products/admin', 'ProductController@adminProductsList');//liste des produits admins
-Route::get('/products/user', 'ProductController@userProductsList');//liste des produits utilisateurs
-Route::get('products/admin/{product}', 'ProductController@adminGetProductById');
-Route::get('products/user/{product}', 'ProductController@userGetProductById');
-Route::post('products', 'ProductController@store');
-Route::put('products/{product}', 'ProductController@update');
-Route::delete('products/{product}', 'ProductController@delete');
+Route::get('products/admin', 'ProductController@adminProductsList');//liste des produits admins
+Route::get('products/user', 'ProductController@userProductsList');//liste des produits utilisateurs
+Route::get('products/admin/{product}', 'ProductController@adminGetProductById');//get porduct as admin
+Route::get('products/user/{product}', 'ProductController@userGetProductById');//get product as user
+Route::post('products', 'ProductController@store');//add product admin only
+Route::put('products/{product}', 'ProductController@update');//update product admin only
+Route::delete('products/{product}', 'ProductController@delete');//delete product admin only
+
+Route::get('categories', 'CategoryController@categoriesList');//liste des produits admins
+Route::post('categories', 'CategoryController@store');//add product admin only
+Route::put('categories/{category}', 'CategoryController@update');//update product admin only
+Route::delete('categories/{category}', 'CategoryController@delete');//delete product admin only
